Implementado upload no editar candidato

+++ web/lib/WebZeta/Controller/Admin/Form/Candidate.pm
@@ -95,17 +95,38 @@ sub process_edit : Chained('base') : PathPart('candidate') : Args(1) {
 		if( $c->req->upload ) {
 			my $upload = $c->req->upload('img_profile');
 			
-			$api->stash_result(
-				$c, 'candidates/upload_file',
-				method => 'UPLOAD',
-				body   => [
-					candidate_id 	=> $c->stash->{id},
-					file 			=> [ $upload->tempname ]
-				]
-			);
+			if($upload) {
+				$api->stash_result(
+					$c, 'candidates/upload_file',
+					method => 'UPLOAD',
+					body   => [
+						candidate_id 	=> $c->stash->{id},
+						file 			=> [ $upload->tempname ],
+						type			=> 'profile'
+					]
+				);
+			
+				if ( $c->stash->{error} ) {
+					$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao associar imagem de perfil ao candidato.' ] );
+				}
+			}
+			
+			$upload = $c->req->upload('gvt_program');
+			
+			if($upload) {
+				$api->stash_result(
+					$c, 'candidates/upload_file',
+					method => 'UPLOAD',
+					body   => [
+						candidate_id 	=> $c->stash->{id},
+						file 			=> [ $upload->tempname ],
+						type			=> 'program'
+					]
+				);
 		
-			if ( $c->stash->{error} ) {
-				$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao associar imagem de perfil ao candidato.' ] );
+				if ( $c->stash->{error} ) {
+					$c->detach( '/form/redirect_error', [ '/admin/candidate/index', {}, 'Problemas ao cadastrar programa de governo ao candidato.' ] );
+				}
 			}
 		}
